refactor(log): split log.ts (382 LOC) into log-types/log-parse/log-session (each ≤200 LOC)
Resolves Gate 4 review finding. log.ts becomes a 3-line barrel re-export.
All 250 tests pass, tsc --noEmit clean.
