+++ aloop/cli/dashboard/src/lib/log-session.ts
@@ -1,0 +1,110 @@
+import type { ManifestPayload, ProviderHealth, SessionSummary } from './log-types';
+import { isRecord, str, numStr } from './log-parse';
+
+export function toSession(source: Record<string, unknown>, fallback: string, isActive: boolean): SessionSummary {
+  return {
+    id: str(source, ['session_id', 'id'], fallback),
+    name: str(source, ['session_id', 'name', 'session_name'], fallback),
+    projectName: str(source, ['project_name'], '') || (() => {
+      const root = str(source, ['project_root'], '');
+      if (root) { const parts = root.replace(/[\\/]+$/, '').split(/[\\/]/); return parts[parts.length - 1] || root; }
+      return fallback.split('-').slice(0, -1).join('-') || fallback;
+    })(),
+    status: str(source, ['state', 'status'], 'unknown'),
+    phase: str(source, ['phase', 'mode'], ''),
+    elapsed: str(source, ['elapsed', 'elapsed_time', 'duration'], '--'),
+    iterations: numStr(source, ['iteration', 'iterations']),
+    isActive,
+    branch: str(source, ['branch'], ''),
+    startedAt: str(source, ['started_at'], ''),
+    endedAt: str(source, ['ended_at'], ''),
+    pid: numStr(source, ['pid'], ''),
+    provider: str(source, ['provider'], ''),
+    workDir: str(source, ['work_dir'], ''),
+    stuckCount: typeof source.stuck_count === 'number' ? source.stuck_count : 0,
+  };
+}
+
+export function latestQaCoverageRefreshSignal(log: string): string | null {
+  if (!log) return null;
+  const lines = log.split('\n');
+  for (let i = lines.length - 1; i >= 0; i--) {
+    const line = lines[i]?.trim();
+    if (!line) continue;
+    try {
+      const entry = JSON.parse(line);
+      if (!isRecord(entry)) continue;
+      const event = str(entry, ['event', 'type']);
+      const phase = str(entry, ['phase', 'mode']).toLowerCase();
+      if (event !== 'iteration_complete' || phase !== 'qa') continue;
+      const timestamp = str(entry, ['timestamp', 'ts', 'time', 'created_at']);
+      const iterationRaw = entry.iteration;
+      const iteration = typeof iterationRaw === 'number' ? String(iterationRaw)
+        : typeof iterationRaw === 'string' ? iterationRaw : '';
+      return `${timestamp}|${iteration}|${line}`;
+    } catch {
+      // Skip non-JSON lines in log stream.
+    }
+  }
+  return null;
+}
+
+export function deriveProviderHealth(log: string, configuredProviders?: string[]): ProviderHealth[] {
+  const providers = new Map<string, ProviderHealth>();
+
+  // Seed configured providers as baseline with 'unknown' status
+  if (configuredProviders) {
+    for (const name of configuredProviders) {
+      if (name) providers.set(name, { name, status: 'unknown', lastEvent: '' });
+    }
+  }
+
+  for (const line of log.split('\n')) {
+    const trimmed = line.trim();
+    if (!trimmed) continue;
+    try {
+      const obj = JSON.parse(trimmed);
+      if (!isRecord(obj)) continue;
+      const event = str(obj, ['event']);
+      const provider = str(obj, ['provider']);
+      const ts = str(obj, ['timestamp']);
+      if (!provider) continue;
+
+      if (event === 'provider_cooldown') {
+        providers.set(provider, {
+          name: provider,
+          status: 'cooldown',
+          lastEvent: ts,
+          reason: str(obj, ['reason']),
+          consecutiveFailures: typeof obj.consecutive_failures === 'number' ? obj.consecutive_failures : undefined,
+          cooldownUntil: str(obj, ['cooldown_until']),
+        });
+      } else if (event === 'provider_recovered') {
+        providers.set(provider, { name: provider, status: 'healthy', lastEvent: ts });
+      } else if (event === 'iteration_complete' || event === 'iteration_error') {
+        if (!providers.has(provider) || providers.get(provider)!.status === 'unknown') {
+          providers.set(provider, { name: provider, status: 'healthy', lastEvent: ts });
+        } else {
+          const existing = providers.get(provider)!;
+          if (event === 'iteration_complete' && existing.status !== 'cooldown') {
+            existing.status = 'healthy';
+          }
+          existing.lastEvent = ts;
+        }
+      }
+    } catch { /* skip */ }
+  }
+  return Array.from(providers.values()).sort((a, b) => a.name.localeCompare(b.name));
+}
+
+export function slugify(text: string): string {
+  return text.toLowerCase().replace(/[^\w\s-]/g, '').replace(/\s+/g, '-').replace(/-+/g, '-').trim();
+}
+
+/** Find iterations (older than current) that have the same artifact path — for history scrubbing */
+export function findBaselineIterations(artifactPath: string, currentIteration: number, allManifests: ManifestPayload[]): number[] {
+  return allManifests
+    .filter((m) => m.iteration < currentIteration && m.artifacts.some((a) => a.path === artifactPath))
+    .map((m) => m.iteration)
+    .sort((a, b) => b - a); // newest first
+}
